requirements.txt
fastapi
uvicorn
oracledb
python-dotenv
email-validator
python-dotenv==1.0.0